puto onedrive

+++ OnlineVideoPlayer/Config.cs
@@ -16,10 +16,8 @@ namespace OnlineVideoPlayer
 
         public static T GetConfig<T>(string keyName, T defaultValue = default)
         {
-            if (!File.Exists(Program.VideoPlayerConfigPath))
-            {
-                return defaultValue;
-            }
+            if (!File.Exists(Program.VideoPlayerConfigPath)) return defaultValue;
+            
 
             try
             {
@@ -44,10 +42,8 @@ namespace OnlineVideoPlayer
 
         public static void SaveConfig(string keyName, object objectData)
         {
-            if (!File.Exists(Program.VideoPlayerConfigPath))
-            {
-                File.WriteAllBytes(Program.VideoPlayerConfigPath, VideoPlayer.Compress(Encoding.UTF8.GetBytes("{\"Test\":\"True\"}")));
-            }
+            if (!File.Exists(Program.VideoPlayerConfigPath)) File.WriteAllBytes(Program.VideoPlayerConfigPath, VideoPlayer.Compress(Encoding.UTF8.GetBytes("{\"test\":\"True\"}")));
+            
 
             try
             {
@@ -63,11 +59,8 @@ namespace OnlineVideoPlayer
             {
                 Console.WriteLine(ex.ToString());
 
-                if (File.Exists(Program.VideoPlayerConfigPath))
-                {
-                    File.Delete(Program.VideoPlayerConfigPath);
-                }
-
+                if (File.Exists(Program.VideoPlayerConfigPath)) File.Delete(Program.VideoPlayerConfigPath);
+                
                 throw ex;
             }
         }
